Extract out a Highlight component for Quick Switch matching, and apply fuzzysort to all fuzzy matching to be consistent
